# fix(perf): use request streams to reduce memory usage