Before this fix, max mem usage scanning nodejs-ai-platform was ~3GB.  After, it was ~0.5GB.

Before:
```
npm run docs-test   513.31s  user 29.49s system 108% cpu 8:19.94 total
avg shared (code):         0 KB
avg unshared (data/stack): 0 KB
total (sum):               0 KB
max memory:                3016600 KB
page faults from disk:     0
other page faults:         9671141
```

After:
```
npm run docs-test   520.45s  user 26.06s system 110% cpu 8:14.96 total
avg shared (code):         0 KB
avg unshared (data/stack): 0 KB
total (sum):               0 KB
max memory:                561248 KB
page faults from disk:     0
other page faults:         7287496
```